moved config.h to includes use AM_CPPFLAGS instead of INCLUDES some small code fixes

+++ src/fun.c
@@ -44,7 +44,7 @@ random_goto()
 {
     char *thegoto;
     int ret;
-    int timescalled;
+    size_t timescalled;
 
     if((rand() % 10) == 1) {
 	thegoto = mono_sql_random_goto();
